// dist/common/constants/validation-messages.constant.d.ts
export declare const AuthValidationMessages: {
    EMAIL_REQUIRED: string;
    EMAIL_VALID: string;
    PASSWORD_REQUIRED: string;
    PASSWORD_STRING: string;
    PASSWORD_LENGTH: string;
    PASSWORD_PATTERN: string;
    PASSWORD_CONFIRM_REQUIRED: string;
    PASSWORD_NOT_MATCH: string;
};
export declare const UserValidationMessages: {
    USERNAME_REQUIRED: string;
    USERNAME_STRING: string;
    USERNAME_LENGTH: string;
    USER_NOT_FOUND: string;
};
export declare const LoginMessages: {
    LOGIN_SUCCESS: string;
    LOGIN_FAILED: string;
};
export declare const RegistrationMessages: {
    REGISTER_SUCCESS: string;
    EMAIL_EXISTS: string;
};
export declare const EmailVerificationMessages: {
    EMAIL_VERIFICATION_SENT: string;
    EMAIL_VERIFIED: string;
    EMAIL_ALREADY_VERIFIED: string;
    EMAIL_VERIFICATION_EXPIRED: string;
    EMAIL_VERIFICATION_INVALID: string;
    EMAIL_VERIFICATION_NOT_FOUND: string;
};
export declare const PasswordMessages: {
    PASSWORD_RESET_SENT: string;
    PASSWORD_RESET_SUCCESS: string;
    PASSWORD_RESET_EXPIRED: string;
    PASSWORD_RESET_INVALID: string;
    PASSWORD_RESET_NOT_FOUND: string;
    PASSWORD_OLD_REQUIRED: string;
    PASSWORD_OLD_STRING: string;
    PASSWORD_NEW_REQUIRED: string;
    PASSWORD_NEW_STRING: string;
    PASSWORD_CONFIRM_REQUIRED: string;
    PASSWORD_NEW_LENGTH: string;
    PASSWORD_NEW_PATTERN: string;
    PASSWORD_CONFIRM_STRING: string;
    PASSWORD_NOT_MATCH: string;
    PASSWORD_SAME_AS_OLD: string;
    PASSWORD_INCORRECT: string;
    PASSWORD_CHANGE_SUCCESS: string;
};
export declare const TokenMessages: {
    TOKEN_REQUIRED: string;
    TOKEN_STRING: string;
    TOKEN_INVALID: string;
};
export declare const EmailErrorMessages: {
    EMAIL_SEND_FAILED: string;
    EMAIL_CONFIG_ERROR: string;
};
export declare const GeneralErrorMessages: {
    INTERNAL_SERVER_ERROR: string;
    UNAUTHORIZED: string;
    FORBIDDEN: string;
};
export declare const ValidationMessages: {
    INTERNAL_SERVER_ERROR: string;
    UNAUTHORIZED: string;
    FORBIDDEN: string;
    EMAIL_SEND_FAILED: string;
    EMAIL_CONFIG_ERROR: string;
    TOKEN_REQUIRED: string;
    TOKEN_STRING: string;
    TOKEN_INVALID: string;
    PASSWORD_RESET_SENT: string;
    PASSWORD_RESET_SUCCESS: string;
    PASSWORD_RESET_EXPIRED: string;
    PASSWORD_RESET_INVALID: string;
    PASSWORD_RESET_NOT_FOUND: string;
    PASSWORD_OLD_REQUIRED: string;
    PASSWORD_OLD_STRING: string;
    PASSWORD_NEW_REQUIRED: string;
    PASSWORD_NEW_STRING: string;
    PASSWORD_CONFIRM_REQUIRED: string;
    PASSWORD_NEW_LENGTH: string;
    PASSWORD_NEW_PATTERN: string;
    PASSWORD_CONFIRM_STRING: string;
    PASSWORD_NOT_MATCH: string;
    PASSWORD_SAME_AS_OLD: string;
    PASSWORD_INCORRECT: string;
    PASSWORD_CHANGE_SUCCESS: string;
    EMAIL_VERIFICATION_SENT: string;
    EMAIL_VERIFIED: string;
    EMAIL_ALREADY_VERIFIED: string;
    EMAIL_VERIFICATION_EXPIRED: string;
    EMAIL_VERIFICATION_INVALID: string;
    EMAIL_VERIFICATION_NOT_FOUND: string;
    REGISTER_SUCCESS: string;
    EMAIL_EXISTS: string;
    LOGIN_SUCCESS: string;
    LOGIN_FAILED: string;
    USERNAME_REQUIRED: string;
    USERNAME_STRING: string;
    USERNAME_LENGTH: string;
    USER_NOT_FOUND: string;
    EMAIL_REQUIRED: string;
    EMAIL_VALID: string;
    PASSWORD_REQUIRED: string;
    PASSWORD_STRING: string;
    PASSWORD_LENGTH: string;
    PASSWORD_PATTERN: string;
};